// frontend/src/api/model/practiceChapter.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Capítulos del nomenclador.
 */
export type PracticeChapter = typeof PracticeChapter[keyof typeof PracticeChapter];


export const PracticeChapter = {
  CONSULTAS: 'CONSULTAS',
  INTERNACION: 'INTERNACION',
  PRACTICAS_ESPECIALIZADAS: 'PRACTICAS_ESPECIALIZADAS',
  CIRUGIA: 'CIRUGIA',
  OBSTETRICIA: 'OBSTETRICIA',
  ANESTESIA: 'ANESTESIA',
  LABORATORIO: 'LABORATORIO',
  DIAGNOSTICO_POR_IMAGENES: 'DIAGNOSTICO_POR_IMAGENES',
  ANATOMIA_PATOLOGICA: 'ANATOMIA_PATOLOGICA',
  HEMOTERAPIA: 'HEMOTERAPIA',
  KINESIOLOGIA: 'KINESIOLOGIA',
  FONOAUDIOLOGIA: 'FONOAUDIOLOGIA',
  SALUD_MENTAL: 'SALUD_MENTAL',
  ODONTOLOGIA: 'ODONTOLOGIA',
  TRASLADOS: 'TRASLADOS',
  OTROS: 'OTROS',
} as const;