// frontend/src/api/model/practiceSetting.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Ámbito en el que la práctica puede realizarse.
 */
export type PracticeSetting = typeof PracticeSetting[keyof typeof PracticeSetting];


export const PracticeSetting = {
  AMBULATORIO: 'AMBULATORIO',
  INTERNACION: 'INTERNACION',
  AMBOS: 'AMBOS',
} as const;