// frontend/src/api/model/practiceType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type PracticeType = typeof PracticeType[keyof typeof PracticeType];


export const PracticeType = {
  CONSULTA: 'CONSULTA',
  PRACTICA: 'PRACTICA',
  CIRUGIA: 'CIRUGIA',
  LABORATORIO: 'LABORATORIO',
  IMAGENES: 'IMAGENES',
  ANESTESIA: 'ANESTESIA',
  INTERNACION: 'INTERNACION',
  MODULO: 'MODULO',
  TRASLADO: 'TRASLADO',
  OTRO: 'OTRO',
} as const;